requirements.txt
django
django-phonenumber-field
psycopg2
djangorestframework
redis
requests
django-bulk-update